Numeric properties form double tim long

+++ Moves.Net/Model/Activity.cs
@@ -22,16 +22,16 @@ namespace Moves.Net.Model
 		public DateTime EndTime { get; set; }
 
 		[JsonProperty("duration")]
-		public double Duration { get; set; }
+		public long Duration { get; set; }
 
 		[JsonProperty("distance")]
-		public double Distance { get; set; }
+		public long Distance { get; set; }
 
 		[JsonProperty("steps")]
-		public double Steps { get; set; }
+		public long Steps { get; set; }
 
 		[JsonProperty("calories")]
-		public double Calories { get; set; }
+		public long Calories { get; set; }
 
 		[JsonProperty("trackPoints")]
 		public Trackpoint[] TrackPoints { get; set; }
